requirements.txt
Flask
Flask-Babel
Flask-Cache
Flask-SQLAlchemy
Flask-Migrate
Flask-MenuBuilder
Flask-WTF
Flask-Principal
WTForms
WTForms-Alchemy
furl
blinker
iso8601
PyGitHub
JenkinsAPI
